requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
filelock==3.16.1
flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.10.0
gunicorn==20.1.0
huggingface-hub==0.26.3
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
packaging==24.2
pdf2image==1.17.0
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
pytesseract==0.3.13
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
torchvision==0.14.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
